requirements.txt
google-adk
litellm
google-genai